+++ front/src/models/entities/address.ts
@@ -1,0 +1,19 @@
+/**
+ * An address.
+ */
+export type Address = {
+  /**
+   * The first line of the address.
+   */
+  firstLine: string;
+
+  /**
+   * The ZIP code.
+   */
+  zipCode: string;
+
+  /**
+   * The city.
+   */
+  city: string;
+};
