+++ front/src/models/entities/company.ts
@@ -1,0 +1,14 @@
+/**
+ * A company.
+ */
+export type Company = {
+  /**
+   * The id of the company.
+   */
+  id: number;
+
+  /**
+   * The name of the company.
+   */
+  name: string;
+};
