+++ front/src/models/entities/jobCategory.ts
@@ -1,0 +1,14 @@
+/**
+ * The category of a job.
+ */
+export type JobCategory = {
+  /**
+   * The id of the job category.
+   */
+  id: number;
+
+  /**
+   * The name of the job category.
+   */
+  category: string;
+};
